Crate rust_git

A version control system implementing a subset features of Git, written in Rust.

Supported commands include:

init: Initialize a new repository

add: Stage files into the index

rm: Remove files from the index or working directory

commit: Record staged changes with a message

branch: Create, delete, or list branches

checkout: Switch branch or commit

merge: Merge two branches

status: Show working tree status

log: View commit history

Support additional features include:

Detailed, beaufitied output (add -v or --verbose)

Operating multiple files or folders in one command

git log, git status and git branch

Out-of-folder operation, can assign a working directory by -p or --path

DETACH HEAD Mode

Operation avaliability check (check for uncommited changes/stages before checkout or merge), can be turned on by -f false or --force false

Behaviors different to Git:

Use self-designed structure and base64 encoding for object storage (binary-INcompatible with Git)

-f or --force is ON BY DEFAULT

-v or --verbose is OFF BY DEFAULT

Modules

commands 🔒

core 🔒

utils 🔒

Structs

Cli 🔒

CommonArgs 🔒

Enums

Commands 🔒

Functions

main 🔒