rust_git::utils
Module hash

Module: hash

Provides hashing utilities used throughout the Git implementation. Currently, it uses SHA-1 as the hashing algorithm to compute object identifiers, consistent with Git’s internal use of SHA-1 for commit, tree, and blob hashes.

Constants

FOLDER_LENGTH

The length (in characters) used for the first level of folder sharding in the object store.

HASH_LENGTH

The length (in characters) of a hexadecimal-encoded SHA-1 hash string.

Functions

sha1

Computes the SHA-1 hash of the provided byte slice and returns the hexadecimal representation as a string.