Crate rust_git

rust_git provides facilities for reading, manipulating and creating Git repositories.

The project is still in its early stages.

Examples

Finding a changeset

The following example finds a changeset and returns its message.

fn get_message_by_commit_id(id: &ObjectId) -> Result<String, GitError> {
    let path = Path::new("/path/to/repo");
    let repo = Repository::open(&path);
    commit::find_one(&id, &repo).map(|commit| commit.message)
}

Modules

blob
Represents Git blob objects.

branch
Represents Git branches.

commit
Represents Git changesets.

commit_filter
A filter that specifies which changesets are queried.

commit_sort_strategy
Specifies in which order changesets are queried.

error
Every rust_git error.

file_status
Every index file status.

meta
Meta information about Git objects.

object_database
Utilities for working with the repository.

object_header
A generic header for every Git object.

object_id
Represents Git identifiers.

object_type
Represents the types of Git objects.

reference
Repository references.

repository
Represents the actual Git repository.

tag
Represents Git tags.

tag_type
Represents the types of Git tags.